Pseudo-randomly assign locations to nodes if needed (SimulatePubnet)
Currently, if a node has no location, we assign Ashburn as it is one of
the most common locations. However, since more than half of the nodes
(171 / 286) do not have locations, this would likely create significantly
more pairs of nodes with zero latency than the actual network.
This change assigns a location based on the distribution of known locations in the data, which
hopefully will lead to a simulation closer to the pubnet.