Add a special case for pod names (stellar.expert => expert)
Currently, both "stellar.org" and "stellar.expert" get the pod name
"stellar-*" which is a bit confusing.
This commit adds a special case for "stellar.expert" and turn that into
"expert". This change is not very elegant, but it's unlikely that we have
to deal with corner cases like that often, so I figured it's reasonable.

We can't use the full name due to label-length limits.