Increase `numAccounts` in `SlowNodesNetwork` mission
The `SlowNodesNetwork` mission is a little flaky because in some cases
the size of the transaction queue can grow past the number of accounts,
causing loadgen to fail. In investigating this issue I discovered that
the ratio of `numAccounts` to `txRate` is quite low relative to many of
the other supercluster missions. This change increases `numAccounts` to
bring the ratio closer to the ratios in other missions, which should
address this mission's flakiness.