Bump simulate pubnet limits
This change allows pods in the SimulatePubnet mission to burst higher,
both in memory and CPU. These limits match the ones we used for the
large scale simulations to simulate the impact of adding more orgs (see
`SimulatePubnetMixedLoadResourceRequirements`), so the cluster should
tolerate the increase well.

I also removed a comment about CPU quotas that I believe is not true
anymore. Our vCPU quota is much higher than 100 these days.