Add theoretical max TPS results for stellar-core v22.2.0
This change adds results for stellar-core v22.2.0, and also adds a new
column to the table containing the commit hash of supercluster used to
run the test. We will add this hash for all new results going forward.
This is one of a few planned changes we are making to the testing
methodology and reporting. The remaining changes will roll out with
results for the next release.