Remove outdated background overlay option
The `--enable-background-overlay` flag sets
`EXPERIMENTAL_BACKGROUND_OVERLAY_PROCESSING=true` in stellar-core.
However, stellar-core 23.0.0 renamed this option. Since the background
overlay is enabled by default now, and the only use of the supercluster
flag was to enable the stellar-core option, this PR removes the flag
entirely (rather than renaming it) as the flag is now redundant.