Add v22 compatibility mode to `--run-for-max-tps`
This change adds a new option `classic-v22-compat` to the
`--run-for-max-tps` flag. This option omits some testing-only
configuration options that were added in v23.0.0 so that supercluster
remains compatible with at least stellar-core v22.3.0. This is important
for theoretical max TPS testing where we like to retest old versions to
establish a baseline before testing new versions.