Improve parallel catchup v2 scripts (#316)
Uniform range generation can take a long time if there are thousands
of ranges that need to be inserted into redis.
The new method creates a file and loads it into redis in one transaction.
In my testing the script took less than 5s to execute from start to finish.

I also improved the worker script by moving post-catchup redis operations
to a transaction and adding simple retry logic. This should improve data
consistency if there is a temporary problem with redis operations.